Refactor: extract router module and slim main.js (#82)
- Create web/js/router.js with PAGES, getPageFromPath(), navigateToPage()
- Move editFavouriteSearch to pages/search.js (imports navigateToPage from router)
- Update pages/favourites.js to import editFavouriteSearch directly from search.js
- Slim main.js to pure orchestration: imports, error handlers, init, service worker
- Add /js/router.js to service worker static asset cache
- Update integration test to reflect new module boundaries
